requirements.txt
Django==3.2.7
psycopg2==2.9.3
coverage==6.3.2
django-heroku
gunicorn
channels==3.0.4
Pillow==9.0.1
python-dateutil==2.8.2
django-twilio==0.13.2
websockets==9.1
daphne
channels-redis
Twisted
asgiref
cloudinary==1.17.0
django-cloudinary-storage==0.2.3
django-utils-six
